+++ src/app/pages/home/components/index.tsx
@@ -1,0 +1,4 @@
+export { BoardHeader } from './BoardHeader';
+export { BoardFooter } from './BoardFooter';
+export { Task } from './Task';
+export { TaskList } from './TaskList';
